+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* `Army`: Added `.maxPoints` and `.withinMaxPoints`.
+
 ## 0.1.0
 
 * Initial stable release.